# code/docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "ScopeIQ API",
  "description": "BOM, site design and deployment scoping: documents, discrepancies, redlines, RFIs, BOM revisions, estimates and workflow. Authenticate with POST /auth/login (mock personas).",
  "version": "1.0.0-poc"
 },
 "paths": {
...
  "/bom/revisions": {
   "get": {
    "tags": [
     "bom"
    ],
    "summary": "Revisions By Status",
    "description": "Approval queues: e.g. IN_REVIEW for the reviewer, SP_REVIEW for the CX SP, SP_AGREED for Ericsson.",
    "operationId": "revisions_by_status_bom_revisions_get",
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/bom/sites/{site_id}/revisions": {
   "get": {
    "tags": [
     "bom"
    ],
    "summary": "Revisions",
    "operationId": "revisions_bom_sites__site_id__revisions_get",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/bom/revisions/{rev_id}": {
   "get": {
    "tags": [
     "bom"
    ],
    "summary": "Revision",
    "operationId": "revision_bom_revisions__rev_id__get",
    "parameters": [
     {
      "name": "rev_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Rev Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/bom/revisions/{rev_id}/lines": {
   "get": {
    "tags": [
     "bom"
    ],
    "summary": "Lines",
    "operationId": "lines_bom_revisions__rev_id__lines_get",
    "parameters": [
     {
      "name": "rev_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Rev Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "bom"
    ],
    "summary": "Add Line",
    "operationId": "add_line_bom_revisions__rev_id__lines_post",
    "parameters": [
     {
      "name": "rev_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Rev Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BomLineAddIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/bom/revisions/{rev_id}/changes": {
   "get": {
    "tags": [
     "bom"
    ],
    "summary": "Changes",
    "operationId": "changes_bom_revisions__rev_id__changes_get",
    "parameters": [
     {
      "name": "rev_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Rev Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/bom/lines/{line_id}": {
   "patch": {
    "tags": [
     "bom"
    ],
    "summary": "Edit Line",
    "operationId": "edit_line_bom_lines__line_id__patch",
    "parameters": [
     {
      "name": "line_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Line Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BomLineEditIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "bom"
    ],
    "summary": "Delete Line",
    "operationId": "delete_line_bom_lines__line_id__delete",
    "parameters": [
     {
      "name": "line_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Line Id"
      }
     },
     {
      "name": "reason_code",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Reason Code"
      }
     },
     {
      "name": "comment",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Comment"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/bom/revisions/{rev_id}/transition": {
   "post": {
    "tags": [
     "bom"
    ],
    "summary": "Transition",
    "operationId": "transition_bom_revisions__rev_id__transition_post",
    "parameters": [
     {
      "name": "rev_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Rev Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TransitionIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/bom/revisions/{rev_id}/export": {
   "get": {
    "tags": [
     "bom"
    ],
    "summary": "Export",
    "operationId": "export_bom_revisions__rev_id__export_get",
    "parameters": [
     {
      "name": "rev_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Rev Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "BomLineAddIn": {
    "properties": {
     "sector": {
      "type": "string",
      "title": "Sector"
     },
     "catalog_key": {
      "type": "string",
      "title": "Catalog Key"
     },
     "design_qty": {
      "type": "number",
      "minimum": 0.0,
      "title": "Design Qty"
     },
     "spare_qty": {
      "type": "number",
      "minimum": 0.0,
      "title": "Spare Qty",
      "default": 0
     },
     "action": {
      "type": "string",
      "title": "Action",
      "default": "Install"
     },
     "reason_code": {
      "type": "string",
      "title": "Reason Code"
     },
     "comment": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Comment"
     }
    },
    "type": "object",
    "required": [
     "sector",
     "catalog_key",
     "design_qty",
     "reason_code"
    ],
    "title": "BomLineAddIn"
   },
   "BomLineEditIn": {
    "properties": {
     "design_qty": {
      "anyOf": [
       {
        "type": "number",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Design Qty"
     },
     "spare_qty": {
      "anyOf": [
       {
        "type": "number",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Spare Qty"
     },
     "reason_code": {
      "type": "string",
      "title": "Reason Code"
     },
     "comment": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Comment"
     }
    },
    "type": "object",
    "required": [
     "reason_code"
    ],
    "title": "BomLineEditIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "TransitionIn": {
    "properties": {
     "action": {
      "type": "string",
      "title": "Action"
     },
     "reason_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reason Code"
     },
     "comment": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Comment"
     }
    },
    "type": "object",
    "required": [
     "action"
    ],
    "title": "TransitionIn"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}